requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
discord.py==2.4.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
numpy==2.2.1
opencv-python-headless==4.11.0.86
packaging==24.2
pillow==11.1.0
propcache==0.2.1
pycares==4.5.0
pycparser==2.22
pytesseract==0.3.13
setuptools==75.8.0
wheel==0.45.1
yarl==1.18.3